Support ⟸ (leftwards leaning double arrow) in clause syntax.
Both the Rust and go implementation are changed
.
In the go implementation, this is changing from double left arrow which was introduced in the last commit but not yet used
anywhere else.

Leftwards leaning double arrow ⟸ is easier to read.

Syntax should make it easier for users to understand the "direction" of a clause. Input of unicode characters may
be clumsy, but this would not be a problem if a formatter
replaced the :- with ⟸.

PiperOrigin-RevId: 732916156